urequests: Expose HTTP response headers.
Section 3.2 of RFC 7230 states that HTTP header names are case-insensitive.
Further, section 3.2.2 states that HTTP header names do not need to be unique
as long as the semantics of the message are retained when concatenating the
corresponding values into a comma separated list.  When non-unique header
names are present, their order is to be preserved in the concatenation.  This
implementation attempts to comply with these requirements.