logging: Add support for custom handlers.
Any custom handlers will be passed a LogRecord instance which has members
and a dict with "levelname", "message" and "name", to be used for creating
a log message.  The handler list is a global singleton so that sub-logging
objects all use the same set of (root) handlers.

The name of the root handler is also changed from None to "root", to match
CPython.

Signed-off-by: Damien George <damien@micropython.org>