hashlib: import pure python hash first to unbreak the hmac module
With the current import strategy that favours the uhashlib
implementation before using the "internal" hashlib version external
modules like "hmac" break because the uhashlib versions lack properties
like `{digest,block}_size, copy()`.

This commit changes the import so that the internal version of the
hash is tried to be imported first and only if there is no internal
version then the uhashlib version is used.