pprint: Add simple and partial implementation for depth of 1.
This change is to provide a simple implementation of pprint to replace the
dummy placeholder implementation.  This implementation should not be too
slow nor take too much space on disk nor memory (which is the point of
MicroPython).

This does not implement the full specification from CPython, but enough for
simple uses.  It simply expands lists, sets, and dictionaries to multiple
lines if there are more than one element.  Useful for simple printing of
lists, dictionaries, and sets to the console, especially when using the
REPL.