aioble/peripheral.py: Reduce the priority of name in the passive scan response.
A number of fields (services, appearance, manufacturer, name) can appear when a
scan is requested. However there is only so much space in the header so, if a
user has configured multiple fields, some may be 'pushed' into the active
scan response which requires additional communication.

name is, arguably, the least useful of the fields but was given the highest
priority and would 'bump' the others if configured. This change simple reduces
the priority of name and makes it less likely to push the more useful fields
into the active scan response.