aioble/device.py: No default timeout for async disconnected() method.
The value for the `timeout_ms` optional argument to
`DeviceConnection.disconnected()` async method is changed from 60000 to
None. This way users awaiting a device disconnection using `await
connection.disconnected()` won't be surprised by a 1 minute timeout. For
instance, in the `temp_sensor.py` usage example I think it's reasonable
to expect the connection to the client to stay open indefinitely.