Added duration_ms to example
I've added duration_ms to the example scanning code, since it's not an optional parameter. Also, I've added some more aggressive parameters so that it has a better chance of working for someone just trying it for the first time, like myself :) And I've removed the if statement for the name, since that does not cause runtime problems, and allows to show more data. Hope it's useful!