Add bare-minimum manifest.py files to bisect, json, keyword, and types.
These four modules seem to have been missing manifest files; I'm unsure
if it's just been an oversight or if there is a reason for it.

I didn't have much success finding what the versions of these modules
may be, nor did I put too much effort into finding the details commented
in some other manifests (original author, other metadata), but I think
that info should also be covered in the code itself. As in the summary,
this is the bare minimum necessary to be able to use a custom manifest
that requires `keyword`, for example. I've tested this with compiling a
Pi Pico firmware with that requirement.