mip: Support relative source file paths in package.json.
This adds the ability for `mip` to load files from _relative_ source
paths given in the `urls` section of `package.json` files.

Previously, the only accepted forms for source paths were
"http://", "https://", or "github:".

These changes allow a source path to be given without one of these prefixes,
in which case the URL of the parent of the `package.json` is prepended.

So instead of someone forking a repository and having to change source lines
that start with
    github:upstream_repository/filename
into
    github:my_repository/filename
they can now just say
    filename

which will survive forking or cloning without change.

This also allows doing `mip.install()` from a `file://` location.
This makes it easy to do a `git clone` of a repository and test
installation of a package from the local filesystem, especially
using the unix port.

Signed-off-by: Ned Konz <ned@productcreationstudio.com>