python-ecosys/requests: Add User-Agent string.
The HTTP User-Agent string is very useful if malfunctioning IoT devices start
putting excessive load on a server.  It allows the server operator to ban
just the malfunctioning devices, not all users.

It's also a SHOULD in the HTTP specification.  Add it, enabled by default,
but with options to override or disable it.