uaiohttpclient: Implement parsing and handling HTTP auth URLs
Allows HTTP Basic Auth URLs being passed, parsed and handled correctly
to authorise against a server expecting that.
This is a simplification, completely ignoring if the server supports or
asks for HTTP Basic Auth. We simply format and set the header if
the URL indicates so.

This also fixes parsing URLs containing a colon (':') /not/ indicating
a custom port (but e.g. the seperation between user and password as part
of the HTTP Basic Auth credentials).

Previously, the lib died ungracefully when parsing an URL containing
HTTP Basic Auth credentials, as it was always expecting at most only one
colon (':') and also the part coming afterwards being neccesarily a
number (port).