github/workflows: Add workflow to run package tests.
All of the runable package tests are run together in the new `tools/ci.sh`
function called `ci_package_tests_run`.  This is added to a new GitHub
workflow to test the packages as part of CI.

Some packages use `unittest` while others use an ad-hoc test script.
Eventually it would be good to unify all the package tests to use
`unittest`.

Signed-off-by: Damien George <damien@micropython.org>