contextlib: Use a list instead of deque for exit callbacks.
Since deque was removed from this repository the built-in one needs to be
used, and that doesn't have unbounded growth.  So use a list instead, which
is adequate becasue contextlib only needs append and pop, not double ended
behaviour (the previous pure-Python implementation of deque that was used
here anyway used a list as its storage container).

Also tweak the excessive-nesting test so it uses less memory and can run on
the unix port.

Signed-off-by: Damien George <damien@micropython.org>