requests: Use the host in the redirect url, not the one in headers.
The host in headers extracted from the original url may not be the same as
the host in the redirect url.  Poping out the host in headers force the
code to use the host in the redirect url, otherwise the redirect may fail
due to inconsistence of hosts in the original url and the redirect url.

Signed-off-by: 黃昕暐 <meebox@gmail.com>