drivers/sensor/mcp9808: Add MCP9808 temperature sensor driver.
Add single-file module for the Microchip MCP9808 precision temperature
sensor. The module provides a class for reading the temperature and for
configuring the sensor. It has type/value checking for some of the
critical parts of the configuration as well as a debug mode for easier
testing during development.

It includes a test file designe to run directly on a board with a sensor
connected to test both the module and the sensor.

Both module and tests file are thoroughly documented in the code
directly. For more information read README.md on MarcoMiano/mip-mcp9808

Signed-off-by: MarcoMiano <58668074+MarcoMiano@users.noreply.github.com>