datetime: Apply localtz patch to include naive date/time support.
This commit applies the existing `localtz.patch` patch to add support for
naive datetime objects.  That is, objects that don't have any info about
the current timezone.

This allows `datetime.datetime.now()` to work; prior to this patch it would
raise NotImplementedError.

Although we don't really have support for localtime vs gmtime on
bare-metal, ports such as the unix port and webassembly port do have this
distinction, and for them being able to do `datetime.datetime.now()` is
quite important (at least, that's what users expect to be able to do).

The associated unittest test has been updated.

This patch changes the size of datetime.mpy: 8466 -> 8897, so +431 bytes.

Signed-off-by: Damien George <damien@micropython.org>